requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
appdirs==1.4.4            # via fs
boto3==1.16.57            # via fs-s3fs
botocore==1.19.57         # via boto3, s3transfer
coverage==5.3.1           # via -r requirements/test.in
ddt==1.4.1                # via -r requirements/test.in
django-pyfs==3.0          # via -r requirements/test.in
edx-opaque-keys==2.1.1    # via -r requirements/test.in
fs-s3fs==1.1.1            # via django-pyfs
fs==2.4.12                # via django-pyfs, fs-s3fs, xblock
jmespath==0.10.0          # via boto3, botocore
lazy==1.4                 # via -r requirements/test.in
lxml==4.6.2               # via xblock
mako==1.1.4               # via xblock-utils
markupsafe==1.1.1         # via mako, xblock
mock==3.0.5               # via -r requirements/test.in
pbr==5.5.1                # via stevedore
pymongo==3.11.2           # via edx-opaque-keys
python-dateutil==2.8.1    # via botocore, xblock
pytz==2020.5              # via django, fs, xblock
pyyaml==5.3.1             # via xblock
s3transfer==0.3.4         # via boto3
simplejson==3.17.2        # via xblock-utils
six==1.15.0               # via edx-opaque-keys, fs, fs-s3fs, mock, python-dateutil, stevedore, xblock
sqlparse==0.4.1           # via django
stevedore==1.32.0         # via edx-opaque-keys
typing==3.7.4.3           # via fs
urllib3==1.26.2           # via botocore
web-fragments==1.0.0      # via xblock, xblock-utils
webob==1.8.6              # via xblock
xblock-sdk==0.2.2         # via -r requirements/test.in
xblock-utils==2.1.2       # via -r requirements/test.in
xblock==1.4.0             # via xblock-utils

# The following packages are considered to be unsafe in a requirements file:
# setuptools
